Format script files and change from i32 to i64

+++ src/tenma_serial/tenma_commands.rs
@@ -1,11 +1,11 @@
 pub enum TenmaCommand {
     ISet {
         channel: u8,
-        current: f32,
+        current: f64,
     },
     VSet {
         channel: u8,
-        voltage: f32,
+        voltage: f64,
     },
     #[allow(dead_code)]
     Beep(bool),
